OBJ Loading with integration of tinyObjLoader
+ refactor for static use of Graphics Handler

+++ Comphi/src/Comphi/Renderer/IObjects/IIndexBuffer.h
@@ -2,7 +2,7 @@
 
 namespace Comphi {
 
-	typedef uint16_t Index;
+	typedef uint32_t Index;
 	typedef std::vector<Index> IndexArray;
 
 	class IIndexBuffer
